requirements.txt
pathlib
feedparser
request